+++ CHANGELOG.md
@@ -22,6 +22,7 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 - Drop the `approved` flag from locallang labels (#745)
 
 ### Fixed
+- Stop using `PATH_site` in TYPO3 9LTS (#780)
 - Stop using the Core-provided whitespace constants (#778)
 - Fix PHPStan level 0 warnings (#775)
 - Add a missing comma in a language label (#771)